backend/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-filter==2.3.0
djangorestframework==3.11.1
gunicorn==20.0.4
numpy==1.19.1
Pillow==7.2.0
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
